--- a/sm2/sm2.go
+++ b/sm2/sm2.go
@@ -198,10 +198,10 @@ func Sm2Verify(pub *PublicKey, msg, uid []byte, r, s *big.Int) bool {
 	return x.Cmp(r) == 0
 }
 
-/* 
+/*
     za, err := ZA(pub, uid)
 	if err != nil {
-		return 
+		return
 	}
 	e, err := msgHash(za, msg)
 	hash=e.getBytes()
@@ -401,7 +401,7 @@ func ZA(pub *PublicKey, uid []byte) ([]byte, error) {
 	if uidLen > 0 {
 		za.Write(uid)
 	}
-	za.Write(sm2P256ToBig(&sm2P256.a).Bytes())
+	za.Write(sm2P256ToBig(sm2P256.a).Bytes())
 	za.Write(sm2P256.B.Bytes())
 	za.Write(sm2P256.Gx.Bytes())
 	za.Write(sm2P256.Gy.Bytes())
@@ -621,7 +621,6 @@ func getLastBit(a *big.Int) uint {
 }
 
 // crypto.Decrypter
-func (priv *PrivateKey) Decrypt(_ io.Reader, msg []byte, _ crypto.DecrypterOpts) (plaintext []byte, err error){
+func (priv *PrivateKey) Decrypt(_ io.Reader, msg []byte, _ crypto.DecrypterOpts) (plaintext []byte, err error) {
 	return Decrypt(priv, msg)
 }
-