--- a/sm2/utils.go
+++ b/sm2/utils.go
@@ -11,14 +11,14 @@ func Decompress(a []byte) *PublicKey {
 	P256Sm2()
 	x := new(big.Int).SetBytes(a[1:])
 	curve := sm2P256
-	sm2P256FromBig(&xx, x)
-	sm2P256Square(&xx3, &xx)       // x3 = x ^ 2
-	sm2P256Mul(&xx3, &xx3, &xx)    // x3 = x ^ 2 * x
-	sm2P256Mul(&aa, &curve.a, &xx) // a = a * x
-	sm2P256Add(&xx3, &xx3, &aa)
-	sm2P256Add(&xx3, &xx3, &curve.b)
+	xx = sm2P256FromBig(xx, x)
+	xx3 = sm2P256Square(xx)      // x3 = x ^ 2
+	xx3 = sm2P256Mul(xx3, xx)    // x3 = x ^ 2 * x
+	aa = sm2P256Mul(curve.a, xx) // a = a * x
+	xx3 = sm2P256Add(xx3, aa)
+	xx3 = sm2P256Add(xx3, curve.b)
 
-	y2 := sm2P256ToBig(&xx3)
+	y2 := sm2P256ToBig(xx3)
 	y := new(big.Int).ModSqrt(y2, sm2P256.P)
 	if getLastBit(y)+2 != uint(a[0]) {
 		y.Sub(sm2P256.P, y)
@@ -41,8 +41,6 @@ func Compress(a *PublicKey) []byte {
 	return buf
 }
 
-
-
 func SignDigitToSignData(r, s *big.Int) ([]byte, error) {
 	return asn1.Marshal(sm2Signature{r, s})
 }
@@ -56,4 +54,3 @@ func SignDataToSignDigit(sign []byte) (*big.Int, *big.Int, error) {
 	}
 	return sm2Sign.R, sm2Sign.S, nil
 }
-