pkg: github.com/Hyperledger-TWGC/tjfoc-gm/sm2
BenchmarkSM2Sign-16                 4286            279977 ns/op            5445 B/op        117 allocs/op
BenchmarkSM2Sign-16                 3723            279568 ns/op            5410 B/op        117 allocs/op
BenchmarkSM2Sign-16                 4162            273926 ns/op            5445 B/op        117 allocs/op
BenchmarkSM2Sign-16                 4351            277503 ns/op            5437 B/op        117 allocs/op
BenchmarkSM2Sign-16                 4088            275262 ns/op            5413 B/op        117 allocs/op
BenchmarkSM2Verify-16                758           1456273 ns/op           75869 B/op       1578 allocs/op
BenchmarkSM2Verify-16                820           1435853 ns/op           78785 B/op       1611 allocs/op
BenchmarkSM2Verify-16                854           1463761 ns/op           79467 B/op       1635 allocs/op
BenchmarkSM2Verify-16                858           1429726 ns/op           76007 B/op       1570 allocs/op
BenchmarkSM2Verify-16                795           1435971 ns/op           76825 B/op       1586 allocs/op
BenchmarkEcdsaSign-16              51531             22001 ns/op            2639 B/op         32 allocs/op
BenchmarkEcdsaSign-16              54784             23047 ns/op            2638 B/op         32 allocs/op
BenchmarkEcdsaSign-16              47797             27182 ns/op            2639 B/op         32 allocs/op
BenchmarkEcdsaSign-16              51690             23858 ns/op            2639 B/op         32 allocs/op
BenchmarkEcdsaSign-16              55077             22102 ns/op            2639 B/op         32 allocs/op
BenchmarkEcdsaVerify-16            18684             64592 ns/op             825 B/op         16 allocs/op
BenchmarkEcdsaVerify-16            18837             63920 ns/op             825 B/op         16 allocs/op
BenchmarkEcdsaVerify-16            18700             63698 ns/op             825 B/op         16 allocs/op
BenchmarkEcdsaVerify-16            18973             63611 ns/op             825 B/op         16 allocs/op
BenchmarkEcdsaVerify-16            18903             64891 ns/op             825 B/op         16 allocs/op
PASS
ok      github.com/Hyperledger-TWGC/tjfoc-gm/sm2        31.372s